+++ ethers-core/src/types/chain.rs
@@ -36,6 +36,8 @@ pub enum Chain {
     BinanceSmartChainTestnet = 97,
     Arbitrum = 42161,
     ArbitrumTestnet = 421611,
+    Cronos = 25,
+    CronosTestnet = 338,
 }
 
 impl fmt::Display for Chain {
@@ -63,6 +65,8 @@ impl fmt::Display for Chain {
             Chain::BinanceSmartChainTestnet => "bsc-testnet",
             Chain::Arbitrum => "arbitrum",
             Chain::ArbitrumTestnet => "arbitrum-testnet",
+            Chain::Cronos => "cronos",
+            Chain::CronosTestnet => "cronos-testnet",
         };
 
         write!(formatter, "{}", chain)
@@ -99,6 +103,8 @@ impl TryFrom<u64> for Chain {
             42 => Chain::Kovan,
             100 => Chain::XDai,
             137 => Chain::Polygon,
+            250 => Chain::Fantom,
+            4002 => Chain::FantomTestnet,
             80001 => Chain::PolygonMumbai,
             43114 => Chain::Avalanche,
             43113 => Chain::AvalancheFuji,
@@ -112,6 +118,8 @@ impl TryFrom<u64> for Chain {
             97 => Chain::BinanceSmartChainTestnet,
             42161 => Chain::Arbitrum,
             421611 => Chain::ArbitrumTestnet,
+            25 => Chain::Cronos,
+            338 => Chain::CronosTestnet,
             _ => return Err(ParseChainError(chain.to_string())),
         })
     }
@@ -143,6 +151,8 @@ impl FromStr for Chain {
             "bsc-testnet" => Chain::BinanceSmartChainTestnet,
             "arbitrum" => Chain::Arbitrum,
             "arbitrum-testnet" => Chain::ArbitrumTestnet,
+            "cronos" => Chain::Cronos,
+            "cronos-testnet" => Chain::CronosTestnet,
             _ => return Err(ParseChainError(chain.to_owned())),
         })
     }
@@ -162,7 +172,8 @@ impl Chain {
                 Chain::BinanceSmartChain |
                 Chain::BinanceSmartChainTestnet |
                 Chain::Arbitrum |
-                Chain::ArbitrumTestnet,
+                Chain::ArbitrumTestnet |
+                Chain::Cronos,
         )
     }
 }